Feat: Delete useless file, change filestructure
