+++ frontend/src/App.tsx
@@ -1,12 +1,25 @@
 import React from 'react';
-import './App.css';
+import LoginPage from 'components/pages/Login';
+import NotFound from 'components/pages/NotFound';
+import OAuthCallbackPage from 'components/pages/OAuthCallbackPage';
+import GlobalThemeProvider from 'styles/GlobalThemeProvider';
+import Test from 'components/pages/TestComp';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 
-function App() {
+const App = () => {
+  console.log('app start2');
   return (
-    <div className="App">
-      Hello
-    </div>
+    <GlobalThemeProvider>
+      <Router>
+        <Switch>
+          <Route exact path="/" component={Test} />
+          <Route exact path="/oauthcallback" component={OAuthCallbackPage} />
+          <Route path="/login" component={LoginPage} />
+          <Route path="*" component={NotFound} />
+        </Switch>
+      </Router>
+    </GlobalThemeProvider>
   );
-}
+};
 
 export default App;
