--- frontend/src/store/configureStore.prod.ts
@@ -1,12 +1,0 @@
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
-import api from '../middleware/api';
-import rootReducer from '../reducers';
-
-const configureStore = (preloadedState) => createStore(
-  rootReducer,
-  preloadedState,
-  applyMiddleware(thunk, api),
-);
-
-export default configureStore;
