--- frontend/src/store/configureStore.ts
@@ -1,1 +1,0 @@
-import { createStore } from 'redux';
