--- frontend/src/utils/cookie.ts
@@ -1,9 +1,0 @@
-export const getCookie = (name: string) => {
-  const value = document.cookie.match(`(^|;) ?${name}=([^;]*)(;|$)`);
-  return value ? value[2] : null;
-};
-
-export const deleteCookie = (name: string) => {
-  const date = new Date();
-  document.cookie = `${name}= ; expires=${date.toUTCString()}; path=/`;
-};
